[FEATURE] Introduce Toctree caption
See also https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#directive-toctree
