+++ packages/guides-restructured-text/src/RestructuredText/Directives/ToctreeDirective.php
@@ -4,11 +4,13 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Guides\RestructuredText\Directives;
 
+use phpDocumentor\Guides\Nodes\InlineCompoundNode;
 use phpDocumentor\Guides\Nodes\Menu\TocNode;
 use phpDocumentor\Guides\Nodes\Node;
 use phpDocumentor\Guides\RestructuredText\Parser\BlockContext;
 use phpDocumentor\Guides\RestructuredText\Parser\Directive;
 use phpDocumentor\Guides\RestructuredText\Parser\DirectiveOption;
+use phpDocumentor\Guides\RestructuredText\Parser\Productions\Rule;
 use phpDocumentor\Guides\RestructuredText\Toc\ToctreeBuilder;
 
 /**
@@ -22,8 +24,11 @@ use phpDocumentor\Guides\RestructuredText\Toc\ToctreeBuilder;
  */
 class ToctreeDirective extends BaseDirective
 {
-    public function __construct(private readonly ToctreeBuilder $toctreeBuilder)
-    {
+    /** @param Rule<InlineCompoundNode> $startingRule */
+    public function __construct(
+        private readonly ToctreeBuilder $toctreeBuilder,
+        private readonly Rule $startingRule,
+    ) {
     }
 
     public function getName(): string
@@ -46,6 +51,14 @@ class ToctreeDirective extends BaseDirective
             $options,
         );
 
-        return (new TocNode($toctreeFiles))->withOptions($this->optionsToArray($options));
+        $tocNode =  (new TocNode($toctreeFiles))->withOptions($this->optionsToArray($options));
+
+        if (isset($options['caption'])) {
+            $blockContextOfCaption = new BlockContext($blockContext->getDocumentParserContext(), (string) $options['caption']->getValue());
+            $inlineNode = $this->startingRule->apply($blockContextOfCaption);
+            $tocNode = $tocNode->withCaption($inlineNode);
+        }
+
+        return $tocNode;
     }
 }
