+++ packages/guides/resources/template/html/body/menu/table-of-content.html.twig
@@ -1,3 +1,6 @@
 <div class="toc">
+    {% if node.caption %}
+        <p class="caption">{{ renderNode(node.caption) }}</p>
+    {% endif -%}
     {% include "body/menu/menu-level.html.twig" %}
 </div>
