+++ packages/guides/src/Nodes/Menu/ContentMenuNode.php
@@ -13,15 +13,11 @@ declare(strict_types=1);
 
 namespace phpDocumentor\Guides\Nodes\Menu;
 
-use phpDocumentor\Guides\Nodes\InlineCompoundNode;
-
 use function is_scalar;
 
 /** @link https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#table-of-contents */
 class ContentMenuNode extends MenuNode
 {
-    private InlineCompoundNode|null $caption;
-
     public function getDepth(): int
     {
         if ($this->hasOption('depth') && is_scalar($this->getOption('depth'))) {
@@ -34,18 +30,5 @@ class ContentMenuNode extends MenuNode
     public function isPageLevelOnly(): bool
     {
         return false;
-    }
-
-    public function getCaption(): InlineCompoundNode|null
-    {
-        return $this->caption;
-    }
-
-    public function withCaption(InlineCompoundNode|null $caption): ContentMenuNode
-    {
-        $that = clone $this;
-        $that->caption = $caption;
-
-        return $that;
     }
 }
