+++ tests/Integration/tests/toctree/toctree-caption/expected/index.html
@@ -1,0 +1,17 @@
+<!-- content start -->
+    <div class="section" id="document-title">
+    <h1>Document Title</h1>
+
+            <div class="toc">
+            <p class="caption"><strong>Table of Contents</strong></p>
+    <ul class="menu-level">
+    <li class="toc-item"><a href="/page1.html#page-1">Page 1</a></li>
+
+    <li class="toc-item"><a href="/page2.html#page-2">Page 2</a></li>
+
+    </ul>
+</div>
+
+    </div>
+
+<!-- content end -->
