+++ tests/Integration/tests/toctree/toctree-caption/input/index.rst
@@ -1,0 +1,9 @@
+Document Title
+==============
+
+.. toctree::
+    :caption: **Table of Contents**
+    :titlesonly:
+
+    page1
+    page2
