+++ tests/Integration/tests/toctree/toctree-caption/input/page1.rst
@@ -1,0 +1,8 @@
+Page 1
+======
+
+Chapter 1
+---------
+
+Chapter 2
+---------
