+++ tests/Integration/tests/toctree/toctree-caption/input/page2.rst
@@ -1,0 +1,4 @@
+Page 2
+======
+
+Lorem Ipsum Dolor
